Fix set_vec() so it works correctly with optimization-after-lifting-to-top.
The old implementation could set vectors to values that are too high when an
element in the middle exceeds the vector space bounds in M.  For example,
if M = 5 5 5 5 and src = 2 7 3 1, then dst would be set to 3 0 3 1 instead
of the correct 3 0 0 0.  Fortunately, this doesn't seem to make a difference
in practice.